Initial refactoring effort: extract methods into modules
